Float -> Double
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@63589 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ retex/renderer-desktop/src/main/java/com/himamis/retex/renderer/desktop/geom/Line2DD.java
@@ -45,7 +45,8 @@ package com.himamis.retex.renderer.desktop.geom;
 
 import java.awt.geom.Line2D;
 
-public class Line2DD extends Line2D.Float implements com.himamis.retex.renderer.share.platform.geom.Line2D {
+public class Line2DD extends Line2D.Double
+		implements com.himamis.retex.renderer.share.platform.geom.Line2D {
 
 	private static final long serialVersionUID = 1L;
